Move absent to function package
